requirements.txt
pymongo>=4.0.0
PyYAML>=6.0
Flask>=2.0.0

